requirements.txt
streamlit==1.31.1
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.24
langchain-openai==0.0.6
mysql-connector-python==8.3.0
langchain-experimental
langchain-google-genai
langchain-google-generativeai
google-generativeai
streamlit-chat
retrying
requests
pandas
frappe-client
python-dotenv
langchain_google_genai
torch